Added validation, DTO, update service method, mapper, PasswordEncoder,

See merge request foxminded-teamconnect-2/tracker!23

+++ src/main/java/com/ua/teamconnect/tracker/controller/PositionController.java
@@ -26,7 +26,7 @@ public class PositionController {
 
     @GetMapping
     @ApiResponseNotFound
-    public List<PositionDto> findAll(@RequestParam(required = false) Long departmentId) {
+    public List<PositionDto> findAll(@RequestParam(required = false) Integer departmentId) {
         return positionService.findAll(departmentId);
     }
 }
