+++ src/main/java/com/ua/teamconnect/tracker/model/exception/DepartmentNotFoundException.java
@@ -4,7 +4,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.web.server.ResponseStatusException;
 
 public class DepartmentNotFoundException extends ResponseStatusException {
-    public DepartmentNotFoundException(Long id) {
+    public DepartmentNotFoundException(Integer id) {
         super(HttpStatus.NOT_FOUND, String.format("Department with id %d is not found", id));
     }
 }
